Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/Create-wrong-scheme.any-expected.txt

CONSOLE MESSAGE: line 25: Wrong url scheme for WebSocket http://localhost:49001/echo

PASS Create WebSocket - Pass a URL with a non ws/wss scheme - SYNTAX_ERR is thrown 
